# openapi.yaml
openapi: 3.0.3
info:
  title: FitLife API
  description: |
    REST API untuk aplikasi FitLife — platform nutrisi dan kesehatan berbasis Next.js.

    ## Autentikasi
    API ini mendukung dua metode autentikasi:
    1. **JWT Bearer Token** (direkomendasikan untuk Flutter/mobile) — kirim token di header `Authorization: Bearer <token>`
    2. **Cookie httpOnly** (untuk web browser) — cookie `token` di-set otomatis setelah login

    Token JWT berlaku selama **7 hari** dan berisi payload: `userId`, `role`, `email`.

    ## Base URL
    - **Lokal:** `http://localhost:3000`
    - **Produksi:** `https://<domain-anda>.vercel.app`
  version: "1.0.0"
  contact:
    name: FitLife Dev Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.vercel.app
    description: Production server (Vercel)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token JWT diperoleh dari response login/register. Gunakan di header:
        `Authorization: Bearer <token>`
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          format: email
          example: "budi@example.com"
        password:
          type: string
          minLength: 8
          example: "password123"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "budi@example.com"
        password:
          type: string
          example: "password123"
    AuthResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login berhasil"
        token:
          type: string
          description: JWT Bearer token untuk dipakai di mobile client
          example: "eyJhbGciOiJIUzI1NiJ9..."
        user:
          $ref: '#/components/schemas/UserBasic'
    GoogleAuthRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Google Access Token dari google_sign_in package Flutter
          example: "ya29.a0AfH..."

    # ── User ──────────────────────────────────────────────
    UserBasic:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          example: "budi@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Budi Santoso"
        username:
          type: string
          nullable: true
          example: "budi_s"
        email:
          type: string
          example: "budi@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        phone:
          type: string
          nullable: true
          example: "08123456789"
        birthdate:
          type: string
          format: date-time
          nullable: true
          example: "1995-06-15T00:00:00.000Z"
        weight:
          type: integer
          nullable: true
          description: Berat badan dalam kg
          example: 70
        height:
          type: integer
          nullable: true
          description: Tinggi badan dalam cm
          example: 170
        photo:
          type: string
          nullable: true
          description: URL foto profil dari Cloudinary
          example: "https://res.cloudinary.com/..."
        google_avatar:
          type: string
          nullable: true
          description: URL avatar dari Google OAuth
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
          nullable: true
...
    Artikel:
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          example: "tips-diet-sehat-1710000000000"
        judul:
          type: string
          example: "Tips Diet Sehat untuk Pemula"
        konten:
          type: string
          example: "Isi artikel..."
        kategori:
          type: string
          example: "Nutrisi"
        gambar_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
    ArtikelRequest:
      type: object
      required: [judul, konten, kategori]
      properties:
        judul:
          type: string
          example: "Tips Diet Sehat untuk Pemula"
        konten:
          type: string
          example: "Isi artikel lengkap..."
        kategori:
          type: string
          example: "Nutrisi"
        gambar_url:
          type: string
          nullable: true
        is_featured:
          type: boolean
          example: false
    PaginatedArtikels:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Artikel'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ── Perhitungan ───────────────────────────────────────
...
    Account:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Budi Santoso"
        email:
          type: string
          example: "budi@example.com"
        phone:
          type: string
          nullable: true
        photo:
          type: string
          nullable: true
        google_avatar:
          type: string
          nullable: true
        is_active:
          type: boolean
          example: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    # ── Dashboard ─────────────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Unauthorized"
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email format"]
            password: ["Password minimal 8 karakter"]
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Registrasi akun baru
      description: |
        Membuat akun baru. Setelah berhasil, akan langsung mengembalikan JWT token
        yang bisa langsung dipakai tanpa harus login ulang.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registrasi berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                message: "Registrasi berhasil"
                token: "eyJhbGciOiJIUzI1NiJ9..."
                user:
                  id: 1
                  name: "Budi Santoso"
                  email: "budi@example.com"
                  role: "user"
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email sudah terdaftar
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email sudah terdaftar."
        '500':
          description: Internal Server Error
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login dengan email & password
      description: |
        Login user. Response berisi JWT token yang harus disimpan oleh Flutter
        di `SharedPreferences` atau `FlutterSecureStorage`.

        Server juga menyimpan token ke cookie httpOnly `token` (untuk web).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                message: "Login berhasil"
                token: "eyJhbGciOiJIUzI1NiJ9..."
                user:
                  id: 1
                  name: "Budi Santoso"
                  email: "budi@example.com"
                  role: "user"
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Email atau password salah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Email atau password salah."
        '500':
          description: Internal Server Error
...
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Cek user yang sedang login
      description: Mengembalikan data user berdasarkan token JWT.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Data user aktif
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized (token tidak ada / expired)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User tidak ditemukan
        '500':
          description: Internal Server Error
  /api/auth/google:
    post:
      tags: [Auth]
      summary: Login dengan Google OAuth
      description: |
        Menerima **Google Access Token** dari Flutter dan memvalidasinya ke Google API.
        Jika user belum terdaftar, akan dibuat akun baru secara otomatis.

        **Flutter:** gunakan package `google_sign_in` untuk mendapatkan `accessToken`,
        lalu kirimkan ke endpoint ini.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
            example:
              token: "ya29.a0AfH6SMB..."
      responses:
        '200':
          description: Login Google berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                message: "Login Berhasil"
                token: "eyJhbGciOiJIUzI1NiJ9..."
                role: "user"
                user:
                  id: 5
                  name: "Ani Rahayu"
                  email: "ani@gmail.com"
                  role: "user"
        '400':
          description: Token Google tidak ditemukan atau tidak valid
        '401':
          description: Token Google tidak valid
        '500':
          description: Internal Server Error

  # ============================================================
  # PROFILE
  # ============================================================
...
  /api/artikels:
    get:
      tags: [Artikels]
      summary: Ambil daftar artikel
      description: |
        Ambil semua artikel. Mendukung filter berdasarkan `kategori`, `slug`, `id`,
        dan `featured`. Juga mendukung **pagination opsional**.
      parameters:
        - name: kategori
          in: query
          description: Filter berdasarkan kategori artikel
          schema:
            type: string
          example: "Nutrisi"
        - name: slug
          in: query
          description: Ambil satu artikel by slug
          schema:
            type: string
          example: "tips-diet-sehat-1710000000000"
        - name: id
          in: query
          description: Ambil satu artikel by ID
          schema:
            type: integer
          example: 1
        - name: featured
          in: query
          description: Filter artikel featured saja
          schema:
            type: boolean
          example: true
        - name: page
          in: query
          description: Nomor halaman (aktifkan pagination)
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Jumlah item per halaman
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Daftar artikel
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Artikel'
                    description: Response tanpa pagination
                  - $ref: '#/components/schemas/PaginatedArtikels'
                    description: Response dengan pagination
        '500':
          description: Internal Server Error

    post:
      tags: [Artikels]
      summary: Tambah artikel baru
      description: |
        Menambahkan artikel baru. Slug di-generate otomatis dari `judul`.

        > ⚠️ **Catatan Keamanan:** Belum dilindungi role admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtikelRequest'
      responses:
        '201':
          description: Artikel berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artikel'
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal Server Error

  # ============================================================
  # PERHITUNGAN BMI
  # ============================================================
...
  /api/accounts:
    get:
      tags: [Accounts]
      summary: Daftar semua akun (Admin)
      description: |
        Mengembalikan daftar semua akun user. Bisa difilter berdasarkan `role`.

        > ⚠️ **Catatan Keamanan:** Belum dilindungi. Sebaiknya hanya diakses admin.
      parameters:
        - name: role
          in: query
          description: Filter berdasarkan role
          schema:
            type: string
            enum: [user, admin]
          example: "user"
      responses:
        '200':
          description: Daftar akun
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '500':
          description: Internal Server Error

  # ============================================================
  # ADMIN DASHBOARD
  # ============================================================
...